requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-mssql-backend==2.8.1
djangorestframework==3.12.1
lxml==4.6.2
mysqlclient==2.0.3
Pillow==7.0.0
pkg-resources==0.0.0
pyasn1==0.4.8
PyMySQL==1.0.2
pyodbc==4.0.30
pysmb==1.2.6
python-docx==0.8.10
pytz==2020.1
sqlparse==0.4.1
xlrd==1.2.0
zebra==0.1.0
zpl==0.1.8
zpllibrary==1.0.3
